requirements.txt
appdirs==1.4.3
Django==1.9
django-crispy-forms==1.6.1
django-formtools==1.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
httplib2==0.9.2
MySQL-python==1.2.5
olefile==0.44
packaging==16.8
Pillow==3.4.2
pyparsing==2.2.0
six==1.10.0
uWSGI==2.0.15
